+fix in progress
+fix in progress

+++ LDB_4lab/DEMO/app/models/user_manager.rb
@@ -23,7 +23,9 @@ class UserManager
 
   def delete_user(email)
     user = User.find_by(email: email)
+    return false unless user
     user.destroy
+    true
   end
 
   # TODO: active project checking will be implemented later
